--- a/.github/actions/next-stats-action/src/prepare/repo-setup.js
+++ b/.github/actions/next-stats-action/src/prepare/repo-setup.js
@@ -3,6 +3,7 @@ const fs = require('fs-extra')
 const exec = require('../util/exec')
 const { remove } = require('fs-extra')
 const logger = require('../util/logger')
+const { mockTrace } = require('../../../../../test/lib/mock-trace')
 const semver = require('semver')
 const execa = require('execa')
 
@@ -54,10 +55,11 @@ module.exports = (actionInfo) => {
         }
       }
     },
-    async linkPackages({ repoDir, nextSwcVersion }) {
+    async linkPackages({ repoDir, nextSwcVersion, span = mockTrace() }) {
       let useTestPack = process.env.NEXT_TEST_PACK
 
       if (useTestPack) {
+        console.log('using test pack')
         execa.sync('pnpm', ['turbo', 'run', 'test-pack'], {
           cwd: repoDir,
           env: { NEXT_SWC_VERSION: nextSwcVersion },
@@ -87,7 +89,7 @@ module.exports = (actionInfo) => {
         })
         return pkgPaths
       } else {
-        // TODO: remove after next stable release (current v13.1.2)
+        console.log('not using test pack')
         const pkgPaths = new Map()
         const pkgDatas = new Map()
         let pkgs
@@ -102,83 +104,88 @@ module.exports = (actionInfo) => {
           throw err
         }
 
-        for (const pkg of pkgs) {
-          const pkgPath = path.join(repoDir, 'packages', pkg)
-          const packedPkgPath = path.join(pkgPath, `${pkg}-packed.tgz`)
+        await span
+          .traceChild('prepare packages for linking')
+          .traceAsyncFn(async () => {
+            for (const pkg of pkgs) {
+              const pkgPath = path.join(repoDir, 'packages', pkg)
+              const packedPkgPath = path.join(pkgPath, `${pkg}-packed.tgz`)
 
-          const pkgDataPath = path.join(pkgPath, 'package.json')
-          if (!fs.existsSync(pkgDataPath)) {
-            require('console').log(`Skipping ${pkgDataPath}`)
-            continue
-          }
-          const pkgData = require(pkgDataPath)
-          const { name } = pkgData
-          pkgDatas.set(name, {
-            pkgDataPath,
-            pkg,
-            pkgPath,
-            pkgData,
-            packedPkgPath,
-          })
-          pkgPaths.set(name, packedPkgPath)
-        }
-
-        for (const pkg of pkgDatas.keys()) {
-          const { pkgDataPath, pkgData } = pkgDatas.get(pkg)
-
-          for (const pkg of pkgDatas.keys()) {
-            const { packedPkgPath } = pkgDatas.get(pkg)
-            if (!pkgData.dependencies || !pkgData.dependencies[pkg]) continue
-            pkgData.dependencies[pkg] = packedPkgPath
-          }
-
-          // make sure native binaries are included in local linking
-          if (pkg === '@next/swc') {
-            if (!pkgData.files) {
-              pkgData.files = []
-            }
-            pkgData.files.push('native/*')
-            require('console').log(
-              'using swc binaries: ',
-              await exec(`ls ${path.join(path.dirname(pkgDataPath), 'native')}`)
-            )
-          }
-
-          if (pkg === 'next') {
-            if (nextSwcVersion) {
-              Object.assign(pkgData.dependencies, {
-                '@next/swc-linux-x64-gnu': nextSwcVersion,
-              })
-            } else {
-              if (pkgDatas.get('@next/swc')) {
-                pkgData.dependencies['@next/swc'] =
-                  pkgDatas.get('@next/swc').packedPkgPath
-              } else {
-                pkgData.files.push('native/*')
+              const pkgDataPath = path.join(pkgPath, 'package.json')
+              if (!fs.existsSync(pkgDataPath)) {
+                require('console').log(`Skipping ${pkgDataPath}`)
+                continue
               }
+              const pkgData = require(pkgDataPath)
+              const { name } = pkgData
+              pkgDatas.set(name, {
+                pkgDataPath,
+                pkg,
+                pkgPath,
+                pkgData,
+                packedPkgPath,
+              })
+              pkgPaths.set(name, packedPkgPath)
             }
-          }
 
-          if (pkgData?.scripts?.prepublishOnly) {
-            // There's a bug in `pnpm pack` where it will run
-            // the prepublishOnly script and that will fail.
-            // See https://github.com/pnpm/pnpm/issues/2941
-            delete pkgData.scripts.prepublishOnly
-          }
+            for (const pkg of pkgDatas.keys()) {
+              const { pkgDataPath, pkgData } = pkgDatas.get(pkg)
 
-          await fs.writeFile(
-            pkgDataPath,
-            JSON.stringify(pkgData, null, 2),
-            'utf8'
-          )
-        }
+              for (const pkg of pkgDatas.keys()) {
+                const { packedPkgPath } = pkgDatas.get(pkg)
+                if (!pkgData.dependencies || !pkgData.dependencies[pkg])
+                  continue
+                pkgData.dependencies[pkg] = packedPkgPath
+              }
 
+              // make sure native binaries are included in local linking
+              if (pkg === '@next/swc') {
+                if (!pkgData.files) {
+                  pkgData.files = []
+                }
+                pkgData.files.push('native/*')
+                require('console').log(
+                  'using swc binaries: ',
+                  await exec(
+                    `ls ${path.join(path.dirname(pkgDataPath), 'native')}`
+                  )
+                )
+              }
+
+              if (pkg === 'next') {
+                if (nextSwcVersion) {
+                  Object.assign(pkgData.dependencies, {
+                    '@next/swc-linux-x64-gnu': nextSwcVersion,
+                  })
+                } else {
+                  if (pkgDatas.get('@next/swc')) {
+                    pkgData.dependencies['@next/swc'] =
+                      pkgDatas.get('@next/swc').packedPkgPath
+                  } else {
+                    pkgData.files.push('native/*')
+                  }
+                }
+              }
+
+              if (pkgData?.scripts?.prepublishOnly) {
+                // There's a bug in `pnpm pack` where it will run
+                // the prepublishOnly script and that will fail.
+                // See https://github.com/pnpm/pnpm/issues/2941
+                delete pkgData.scripts.prepublishOnly
+              }
+
+              await fs.writeFile(
+                pkgDataPath,
+                JSON.stringify(pkgData, null, 2),
+                'utf8'
+              )
+            }
+          })
         // wait to pack packages until after dependency paths have been updated
         // to the correct versions
         await Promise.all(
           Array.from(pkgDatas.keys()).map(async (pkgName) => {
-            const { pkg, pkgPath, pkgData, packedPkgPath } =
-              pkgDatas.get(pkgName)
+            const { pkgPath, pkgData, packedPkgPath } = pkgDatas.get(pkgName)
             // Copied from pnpm source: https://github.com/pnpm/pnpm/blob/5a5512f14c47f4778b8d2b6d957fb12c7ef40127/releasing/plugin-commands-publishing/src/pack.ts#L96
             const tmpTarball = path.join(
               pkgPath,
@@ -186,10 +193,15 @@ module.exports = (actionInfo) => {
                 pkgData.version
               }.tgz`
             )
-            await execa('pnpm', ['pack'], {
-              cwd: pkgPath,
-            })
-            await fs.copyFile(tmpTarball, packedPkgPath)
+
+            await span
+              .traceChild(`pack '${pkgName}'`)
+              .traceAsyncFn(async () => {
+                await execa('pnpm', ['pack'], {
+                  cwd: pkgPath,
+                })
+                await fs.copyFile(tmpTarball, packedPkgPath)
+              })
           })
         )