--- a/.github/actions/next-stats-action/src/prepare/repo-setup.js
+++ b/.github/actions/next-stats-action/src/prepare/repo-setup.js
@@ -55,40 +55,157 @@ module.exports = (actionInfo) => {
       }
     },
     async linkPackages({ repoDir, nextSwcVersion }) {
-      const repoDirPkg = require(path.join(repoDir, 'package.json'))
-      const hasTestPackAll = Boolean(repoDirPkg.scripts['test-pack-all'])
-      if (!hasTestPackAll) {
-        // This code will be executed only when running comparison stats against latest stable.
-        // This code is not thread safe (that's what test-pack-all does), but it should be fine here, since main tests won't use this code path.
-        // TODO: remove this code when we release new stable (this was added right after 13.3.0)
-        execa.sync('pnpm', ['turbo', 'run', 'test-pack'], {
-          cwd: repoDir,
-          env: { NEXT_SWC_VERSION: nextSwcVersion },
-        })
-      } else {
-        execa.sync('pnpm', ['test-pack-all'], {
-          cwd: repoDir,
-          env: { NEXT_SWC_VERSION: nextSwcVersion },
-        })
-      }
-      const pkgPaths = new Map()
-      const pkgs = (await fs.readdir(path.join(repoDir, 'packages'))).filter(
-        (item) => !item.startsWith('.')
-      )
+      const useTurbo = Boolean(process.env.NEXT_TEST_PACK)
 
-      pkgs.forEach((pkgDirname) => {
-        const { name } = require(path.join(
-          repoDir,
-          'packages',
-          pkgDirname,
-          'package.json'
-        ))
-        pkgPaths.set(
-          name,
-          path.join(repoDir, 'packages', pkgDirname, `packed-${pkgDirname}.tgz`)
+      if (useTurbo) {
+        const repoDirPkg = require(path.join(repoDir, 'package.json'))
+        const hasTestPackAll = Boolean(repoDirPkg.scripts['test-pack-all'])
+        if (!hasTestPackAll) {
+          // This code will be executed only when running comparison stats against latest stable.
+          // This code is not thread safe (that's what test-pack-all does), but it should be fine here, since main tests won't use this code path.
+          // TODO: remove this code when we release new stable (this was added right after 13.3.0)
+          execa.sync('pnpm', ['turbo', 'run', 'test-pack'], {
+            cwd: repoDir,
+            env: { NEXT_SWC_VERSION: nextSwcVersion },
+          })
+        } else {
+          execa.sync('pnpm', ['test-pack-all'], {
+            cwd: repoDir,
+            env: { NEXT_SWC_VERSION: nextSwcVersion },
+          })
+        }
+        const pkgPaths = new Map()
+        const pkgs = (await fs.readdir(path.join(repoDir, 'packages'))).filter(
+          (item) => !item.startsWith('.')
         )
-      })
-      return pkgPaths
+
+        pkgs.forEach((pkgDirname) => {
+          const { name } = require(path.join(
+            repoDir,
+            'packages',
+            pkgDirname,
+            'package.json'
+          ))
+          pkgPaths.set(
+            name,
+            path.join(
+              repoDir,
+              'packages',
+              pkgDirname,
+              `packed-${pkgDirname}.tgz`
+            )
+          )
+        })
+        return pkgPaths
+      } else {
+        // TODO: remove after next stable release (current v13.1.2)
+        const pkgPaths = new Map()
+        const pkgDatas = new Map()
+        let pkgs
+
+        try {
+          pkgs = await fs.readdir(path.join(repoDir, 'packages'))
+        } catch (err) {
+          if (err.code === 'ENOENT') {
+            require('console').log('no packages to link')
+            return pkgPaths
+          }
+          throw err
+        }
+
+        for (const pkg of pkgs) {
+          const pkgPath = path.join(repoDir, 'packages', pkg)
+          const packedPkgPath = path.join(pkgPath, `${pkg}-packed.tgz`)
+
+          const pkgDataPath = path.join(pkgPath, 'package.json')
+          if (!fs.existsSync(pkgDataPath)) {
+            require('console').log(`Skipping ${pkgDataPath}`)
+            continue
+          }
+          const pkgData = require(pkgDataPath)
+          const { name } = pkgData
+          pkgDatas.set(name, {
+            pkgDataPath,
+            pkg,
+            pkgPath,
+            pkgData,
+            packedPkgPath,
+          })
+          pkgPaths.set(name, packedPkgPath)
+        }
+
+        for (const pkg of pkgDatas.keys()) {
+          const { pkgDataPath, pkgData } = pkgDatas.get(pkg)
+
+          for (const pkg of pkgDatas.keys()) {
+            const { packedPkgPath } = pkgDatas.get(pkg)
+            if (!pkgData.dependencies || !pkgData.dependencies[pkg]) continue
+            pkgData.dependencies[pkg] = packedPkgPath
+          }
+
+          // make sure native binaries are included in local linking
+          if (pkg === '@next/swc') {
+            if (!pkgData.files) {
+              pkgData.files = []
+            }
+            pkgData.files.push('native/*')
+            require('console').log(
+              'using swc binaries: ',
+              await exec(`ls ${path.join(path.dirname(pkgDataPath), 'native')}`)
+            )
+          }
+
+          if (pkg === 'next') {
+            if (nextSwcVersion) {
+              Object.assign(pkgData.dependencies, {
+                '@next/swc-linux-x64-gnu': nextSwcVersion,
+              })
+            } else {
+              if (pkgDatas.get('@next/swc')) {
+                pkgData.dependencies['@next/swc'] =
+                  pkgDatas.get('@next/swc').packedPkgPath
+              } else {
+                pkgData.files.push('native/*')
+              }
+            }
+          }
+
+          if (pkgData?.scripts?.prepublishOnly) {
+            // There's a bug in `pnpm pack` where it will run
+            // the prepublishOnly script and that will fail.
+            // See https://github.com/pnpm/pnpm/issues/2941
+            delete pkgData.scripts.prepublishOnly
+          }
+
+          await fs.writeFile(
+            pkgDataPath,
+            JSON.stringify(pkgData, null, 2),
+            'utf8'
+          )
+        }
+
+        // wait to pack packages until after dependency paths have been updated
+        // to the correct versions
+        await Promise.all(
+          Array.from(pkgDatas.keys()).map(async (pkgName) => {
+            const { pkg, pkgPath, pkgData, packedPkgPath } =
+              pkgDatas.get(pkgName)
+            // Copied from pnpm source: https://github.com/pnpm/pnpm/blob/5a5512f14c47f4778b8d2b6d957fb12c7ef40127/releasing/plugin-commands-publishing/src/pack.ts#L96
+            const tmpTarball = path.join(
+              pkgPath,
+              `${pkgData.name.replace('@', '').replace('/', '-')}-${
+                pkgData.version
+              }.tgz`
+            )
+            await execa('pnpm', ['pack'], {
+              cwd: pkgPath,
+            })
+            await fs.copyFile(tmpTarball, packedPkgPath)
+          })
+        )
+
+        return pkgPaths
+      }
     },
   }
 }