--- a/run-tests.js
+++ b/run-tests.js
@@ -51,9 +51,6 @@ const cleanUpAndExit = async (code) => {
   if (process.env.NEXT_TEST_STARTER) {
     await fs.remove(process.env.NEXT_TEST_STARTER)
   }
-  if (process.env.NEXT_TEST_TEMP_REPO) {
-    await fs.remove(process.env.NEXT_TEST_TEMP_REPO)
-  }
   console.log(`exiting with code ${code}`)
 
   setTimeout(() => {
@@ -256,23 +253,14 @@ async function main() {
     // to avoid having to run yarn each time
     console.log('Creating Next.js install for isolated tests')
     const reactVersion = process.env.NEXT_TEST_REACT_VERSION || 'latest'
-    const { installDir, pkgPaths, tmpRepoDir } = await createNextInstall({
+    const testStarter = await createNextInstall({
       parentSpan: mockTrace(),
       dependencies: {
         react: reactVersion,
         'react-dom': reactVersion,
       },
-      keepRepoDir: true,
     })
-
-    const serializedPkgPaths = []
-
-    for (const key of pkgPaths.keys()) {
-      serializedPkgPaths.push([key, pkgPaths.get(key)])
-    }
-    process.env.NEXT_TEST_PKG_PATHS = JSON.stringify(serializedPkgPaths)
-    process.env.NEXT_TEST_TEMP_REPO = tmpRepoDir
-    process.env.NEXT_TEST_STARTER = installDir
+    process.env.NEXT_TEST_STARTER = testStarter
   }
 
   const sema = new Sema(concurrency, { capacity: testNames.length })
@@ -535,9 +523,7 @@ async function main() {
   }
 }
 
-main()
-  .then(() => cleanUpAndExit(0))
-  .catch((err) => {
-    console.error(err)
-    cleanUpAndExit(1)
-  })
+main().catch((err) => {
+  console.error(err)
+  cleanUpAndExit(1)
+})