… contributing/core/testing.md
… package.json
… scripts/test-pack-package.mts
… test/lib/create-next-install.js
… test/lib/mock-trace.js
… turbo.json