All changes are hidden behind `NEXT_TEST_PACK=1` flag. I want to test this out properly before using it default. Caching issues are hard to catch.

This PR fixes a few issues that @ijjk found in the `test-pack` before he disabled it in https://github.com/vercel/next.js/pull/45602.

There were two issues:
- concurrency issues with `test-pack`
  - I added `test-pack-all`, which uses a filesystem lock to prevent race conditions when running turbo
- `pnpm` cache
  - I haven't observed any issues with `pnpm` cache itself.
	- It's hidden behind a feature flag, so it should be safe to merge and properly test this later	
  - I think this was caused by a bug in `test-pack` turbo config. It didn't depend on `dist/**` so you could cache the wrong output by running the test without running `pnpm build / dev` first.
    - I fixed this by force depending on `dist/**` in turbo. Luckily `pnpm build` is deterministic, so running it won't invalidate packages that didn't change.
  - I'll need to verify this with others to be sure I'm not introducing any cache-related issues that would be hard to debug

## Speed gains summary

Currently, the slowest part of our test is `linkPackages`, which runs `npm pack` on every package in `packages/`. We then install those `.tgz` in our isolated test directory.
Timings of packing:
```
├─ linkPackages 716 ms
│  │  │  ├─ pack '@next/eslint-plugin-next' 431 ms
│  │  │  ├─ pack '@next/font' 638 ms
│  │  │  ├─ pack 'next' 3.2s
│  │  │  ├─ pack '@next/bundle-analyzer' 508 ms
│  │  │  ├─ pack '@next/codemod' 488 ms
│  │  │  ├─ pack '@next/env' 562 ms
│  │  │  ├─ pack '@next/mdx' 550 ms
│  │  │  ├─ pack '@next/plugin-storybook' 452 ms
│  │  │  ├─ pack '@next/polyfill-module' 481 ms
│  │  │  ├─ pack '@next/polyfill-nomodule' 469 ms
│  │  │  ├─ pack '@next/swc' 7.4s
│  │  │  ├─ pack '@next/react-dev-overlay' 570 ms
│  │  │  └─ pack '@next/react-refresh-utils' 508 ms
```
We pack in parallel, so we care only about the maximum, that is `@next/swc`.
Caching with turbo can help us here since we usually don't touch `@next/swc`.

This caching saves us around `~4s` since the second slowest package is `next`, which is way faster to pack. We do some copying, and turbo has a small overhead, so we save `~3s` here.

We also don't need to copy the whole `packages/` dir into temporary directory when running tests. That saves another `1-2s`.

`pnpm` can cache more efficiently since we don't change the paths of `.tgz` files. Which leads to another `1-2s` savings, they do check hash of that file, so it should be safe.

### Final numbers

Jest reports `<8s` when there are no changes in `packages/`, but you can change the test. (13s saved)

When you force `test-pack` on all packages (changing branches), it's slower at `25s` due to the overhead that we have when starting. (4s slower)

When there is a change in `packages/next` (and no other `packages/`), we get these timings:

```
Jest reports 14.5s (6s faster)

createNext 9.8s
├─ init next dev instance 127 µs
├─ createTestDir 9.4s
│  ├─ createNextInstall 9.4s
│  │  ├─ ensure swc binary 6.2 ms
│  │  ├─ linkPackages 6.8s
│  │  └─ run generic install command 2.5s
│  └─ writeInitialFiles 14 ms
└─ start next instance 430 ms
```

Timings from current canary:
```
Jest reports 21s

createNext 17s
├─ init next dev instance 138 µs
├─ createTestDir 16s
│  ├─ createNextInstall 16s
│  │  ├─ ensure swc binary 4.2 ms
│  │  ├─ copy package.json to temp dir 910 µs
│  │  ├─ copy packages to temp dir 2.5s
│  │  ├─ linkPackages 9.5s
│  │  └─ run generic install command 4.1s
│  └─ writeInitialFiles 32 ms
└─ start next instance 365 ms
```

You can run these benchmarks using this command:
```
 rm -f test/.trace/trace && NEXT_TEST_TRACE=1 NEXT_TEST_PACK=1 pnpm test-dev test/e2e/test-utils-tests/basic/basic.test.ts && pnpm node scripts/trace-to-tree.mjs test/.trace/trace
```

## What's next?

- We could reduce `linkPackages` to `~1s` if we could modify the compression level that pnpm uses. `gzip` has `1-9`, and they use `6` (default from node). `1` is really fast. I created [issue](https://github.com/pnpm/pnpm/issues/6393 in pnpm repo, but I probably won't get addressed.

- Another area that could be improved is next itself. `dev` in an empty project currently takes `5s` for the first request (the missing `5s` that can be seen in the traces above). `build` on an empty project takes `15s`.
	- Improvements here are way more valuable

- Jest itself adds like `1.5s` overhead which could be a pretty easy win if we manage to get this down (measured by running an empty test).
	- migrating our e2e test to something faster (like `vitest`) shouldn't be that hard since our tests execute just `createNext`, few fetches, and `playwright`.

fix NEXT-1002